// src/app/services/clients/recyclingbank.client.ts
export interface RecyclingBank {
    id: number;
    plastic: boolean;
    paper: boolean;
    whiteGlass: boolean;
    colouredGlass: boolean;
    metal: boolean;
    capacity: number;
    position: string;
}

export interface RecyclingBankCreateRequest {
    plastic: boolean;
    paper: boolean;
    whiteGlass: boolean;
    colouredGlass: boolean;
    metal: boolean;
    capacity: number;
    position: string;
}

export interface RecyclingBankRequest {
    coordinate?: string | undefined;
    radius: number;
}