// src/app/services/clients/like.client.ts
export interface Like {
    id?: number;
    recyclingBankId?: number;
    profileId?: string | undefined;
}